backend/requirements.txt
# PassiveGuard Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML/Data Science
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.1

# Database (optional - for token storage)
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
structlog==24.1.0
